requirements.txt
astroid==3.3.10
asttokens==3.0.0
autoflake==2.3.1
autopep8==2.3.2
backports.tarfile==1.2.0
black==25.1.0
blinker==1.9.0
bokeh==3.7.3
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.2
cryptography==45.0.4
cycler==0.12.1
decorator==5.2.1
defusedxml==0.7.1
dill==0.4.0
et_xmlfile==2.0.0
executing==2.2.0
Flask==3.1.1
fonttools==4.58.4
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
ipython==9.3.0
ipython_pygments_lexers==1.1.1
isort==6.0.1
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.2.1
jdcal==1.4.1
jedi==0.19.2
Jinja2==3.1.6
jira==3.8.0
jira-agile-metrics==0.24
keyring==25.6.0
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mccabe==0.7.0
memory-profiler==0.61.0
mock==5.2.0
more-itertools==10.7.0
mypy_extensions==1.1.0
narwhals==1.44.0
numpy==2.3.1
oauthlib==3.3.0
openpyxl==3.1.5
packaging==25.0
pandas==2.3.0
parso==0.8.4
pathspec==0.12.1
patsy==1.0.1
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
pluggy==1.6.0
prompt_toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycodestyle==2.14.0
pycparser==2.22
pydicti==1.2.1
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
pylint==3.3.7
pyparsing==3.2.3
pytest==8.4.1
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
ruff==0.12.0
scipy==1.15.3
seaborn==0.13.2
six==1.17.0
stack-data==0.6.3
statsmodels==0.14.4
tlslite==0.4.9
tomlkit==0.13.3
tornado==6.5.1
traitlets==5.14.3
trello==0.9.7.3
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
Werkzeug==3.1.3
xyzservices==2025.4.0
zipp==3.23.0
